refactor: consolidate payment and banking string rules
