--- pkg/rules/string_format_payment_banking.go
@@ -1,158 +1,0 @@
-package rules
-
-import (
-	"github.com/nobl9/govy/internal/messagetemplates"
-	"github.com/nobl9/govy/pkg/govy"
-)
-
-// StringCreditCard ensures the property's value is a plausible digit-only
-// payment card number. It requires a 13- to 19-digit value that passes the
-// Luhn checksum and rejects all-same-digit values.
-func StringCreditCard() govy.Rule[string] {
-	tpl := messagetemplates.Get(messagetemplates.StringCreditCardTemplate)
-
-	return govy.NewRule(func(s string) error {
-		if !isValidCreditCard(s) {
-			return govy.NewRuleErrorTemplate(govy.TemplateVars{
-				PropertyValue: s,
-			})
-		}
-		return nil
-	}).
-		WithErrorCode(ErrorCodeStringCreditCard).
-		WithMessageTemplate(tpl).
-		WithDescription(mustExecuteTemplate(tpl, govy.TemplateVars{}))
-}
-
-// StringLuhnChecksum ensures the property's value is a digit-only string that
-// passes the Luhn checksum.
-func StringLuhnChecksum() govy.Rule[string] {
-	tpl := messagetemplates.Get(messagetemplates.StringLuhnChecksumTemplate)
-
-	return govy.NewRule(func(s string) error {
-		if !isLuhnChecksumValid(s) {
-			return govy.NewRuleErrorTemplate(govy.TemplateVars{
-				PropertyValue: s,
-			})
-		}
-		return nil
-	}).
-		WithErrorCode(ErrorCodeStringLuhnChecksum).
-		WithMessageTemplate(tpl).
-		WithDescription(mustExecuteTemplate(tpl, govy.TemplateVars{}))
-}
-
-// StringBIC ensures the property's value matches the current Business
-// Identifier Code (BIC) syntax.
-func StringBIC() govy.Rule[string] {
-	tpl := messagetemplates.Get(messagetemplates.StringBICTemplate)
-
-	return govy.NewRule(func(s string) error {
-		if !isValidBIC(s) {
-			return govy.NewRuleErrorTemplate(govy.TemplateVars{
-				PropertyValue: s,
-			})
-		}
-		return nil
-	}).
-		WithErrorCode(ErrorCodeStringBIC).
-		WithMessageTemplate(tpl).
-		WithDescription(mustExecuteTemplate(tpl, govy.TemplateVars{}))
-}
-
-// StringBICISO93622014 ensures the property's value matches the ISO 9362:2014
-// Business Identifier Code (BIC) syntax.
-func StringBICISO93622014() govy.Rule[string] {
-	tpl := messagetemplates.Get(messagetemplates.StringBICISO93622014Template)
-
-	return govy.NewRule(func(s string) error {
-		if !isValidBICISO93622014(s) {
-			return govy.NewRuleErrorTemplate(govy.TemplateVars{
-				PropertyValue: s,
-			})
-		}
-		return nil
-	}).
-		WithErrorCode(ErrorCodeStringBICISO93622014).
-		WithMessageTemplate(tpl).
-		WithDescription(mustExecuteTemplate(tpl, govy.TemplateVars{}))
-}
-
-func isValidCreditCard(s string) bool {
-	if len(s) < 13 || len(s) > 19 {
-		return false
-	}
-	if allSameDigit(s) {
-		return false
-	}
-	return isLuhnChecksumValid(s)
-}
-
-func isLuhnChecksumValid(s string) bool {
-	if s == "" {
-		return false
-	}
-
-	sum := 0
-	shouldDouble := false
-	for i := len(s) - 1; i >= 0; i-- {
-		digit := s[i] - '0'
-		if digit > 9 {
-			return false
-		}
-		if shouldDouble {
-			digit *= 2
-			if digit > 9 {
-				digit -= 9
-			}
-		}
-		sum += int(digit)
-		shouldDouble = !shouldDouble
-	}
-	return sum%10 == 0
-}
-
-func allSameDigit(s string) bool {
-	if s == "" {
-		return false
-	}
-	first := s[0]
-	for i := 1; i < len(s); i++ {
-		if s[i] != first {
-			return false
-		}
-	}
-	return true
-}
-
-func isValidBIC(s string) bool {
-	return isValidBICCode(s)
-}
-
-func isValidBICISO93622014(s string) bool {
-	return isValidBICCode(s)
-}
-
-func isValidBICCode(s string) bool {
-	if len(s) != 8 && len(s) != 11 {
-		return false
-	}
-	if !isValidBICCountryCode(s[4:6]) {
-		return false
-	}
-	for i := range 4 {
-		if !isUpperASCIIAlphanumeric(s[i]) {
-			return false
-		}
-	}
-	for i := 6; i < len(s); i++ {
-		if !isUpperASCIIAlphanumeric(s[i]) {
-			return false
-		}
-	}
-	return true
-}
-
-func isUpperASCIIAlphanumeric(b byte) bool {
-	return b >= 'A' && b <= 'Z' || b >= '0' && b <= '9'
-}
